Added Details View
Updated the backend models to incorporate a choice between 2 different
views. The first view will use EquipementType and ItemDetails classes,
which will allow each item to refer to it's respective type. The second
view uses TypeDetails and EquipmentItem which allows for the type to
store a list of Items that are of that type.

No major updates in the equipment service needed other than changing the
class name of the type lists.