Update user model and doc for agreement status
The User model has been given an additional field to track the status
of their agreement, and the design doc has been updated to have a route
to update this field. The previously agreed upon route to retrieve the
user agreement has been removed in favor of a static document stored in
the frontend, similar to the "About the XL" section of the current
website (`frontend/src/app/about/about.component.html`).