Student user agreement backend (#22)
* Edit design-doc

* Complete models

* Complete get equipment availability by type backend.

* Added Details View

Updated the backend models to incorporate a choice between 2 different
views. The first view will use EquipementType and ItemDetails classes,
which will allow each item to refer to it's respective type. The second
view uses TypeDetails and EquipmentItem which allows for the type to
store a list of Items that are of that type.

No major updates in the equipment service needed other than changing the
class name of the type lists.

* Update user model and doc for agreement status

The User model has been given an additional field to track the status
of their agreement, and the design doc has been updated to have a route
to update this field. The previously agreed upon route to retrieve the
user agreement has been removed in favor of a static document stored in
the frontend, similar to the "About the XL" section of the current
website (`frontend/src/app/about/about.component.html`).

* Add num_available and image_url to EquipmentType and update_availibality on request.

* Update User schema and add update_agreement_status

The user schema has been updated ot go along with the user model
changes, and the `PUT` update_agreement_status route has been added to
update the user's agreement status to true when they hit the agree
button.

Note: Make sure to run the script to reset the database after you have
changed the schema.

* Change user_svc to user_service, update agreement wireframe

---------

Co-authored-by: VitorInserra <inserra.vitor@gmail.com>
Co-authored-by: Andrew Lockard <alockard@unc.edu>
Co-authored-by: Vitor Inserra <80555915+VitorInserra@users.noreply.github.com>
Co-authored-by: Jacob Fessler <jacobdan@ad.unc.edu>