Incorporated Benjamin's Suggestions
Updated the Frontend agreement component's typescript to ensure that the
update_agreement api route was called. Also removed the subsequent call
to get_agreement_status as these are concurrent processes. Instead I
just used the return value of update_agreement_status.

Also added documentation to the API routes and began the creation of the
specification file.