Agreement frontend (#37)
* Add and Delete Operating Hours from Coworking Database via API (#136)

* Fix some typing issues in backend/main.py
* Add API for listing Coworking Operating Hours
* Enforce permissions on OperatingHours Create
* Delete OperatingHours via the API

* Add Code File Labels and some Clarification to Event Binding Section (#139)

* Add Clarification to Event Binding (#140)

* Improve documentation of database and branching practices (#138)

* Document Database concerns
* Document Branching Practices and PR/CR Strategies

* Add Extra Resource to Further Reading Section (#141)

* Finalize notes on Merging a PR via Squash and Merge

* Add PR Template to the Docs Page (#142)

This pull request adds a template for PRs to the `branches` docs page.

* Update widgets.md (#143)

* Add Clarification Pt based on a student’s feedback (#144)

* Make Organization, Event, and About Page Public (#146)

* Feature Permissions Document (#149)

This commit adds a document to the docs directory that defines the
standard permissions used in the backend to authorize users for feature
actions. This ensures that permissions are added properly in the
frontend by providing a reference file for existing action/resource
sets.

* Resolve Date Time Issues (#150)

* Add full frontend to backend functionality

* Add css styling

* Incorporated Benjamin's Suggestions

Updated the Frontend agreement component's typescript to ensure that the
update_agreement api route was called. Also removed the subsequent call
to get_agreement_status as these are concurrent processes. Instead I
just used the return value of update_agreement_status.

Also added documentation to the API routes and began the creation of the
specification file.

---------

Co-authored-by: Kris Jordan <kris@cs.unc.edu>
Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>
Co-authored-by: Jade Keegan <97476936+jadekeegan@users.noreply.github.com>
Co-authored-by: Benjamin Eldridge <beldridge@unc.edu>
Co-authored-by: Andrew Lockard <andrew.lockard15@gmail.com>