Update admin equipment user edit component
Right now these components are using dummy data, the get all equipments
route should be updated to provide counts for the base component and edit
component.  We also need to update the frontend's EquipmentType interface.
We might need to use a resolver to pass the values of the EquipmentType
we are editing from the AdminEquipmentBaseComponent to the AdminEquipmentEditComponent
like was done for the organizations editor component.