Added update_item_availablity
Created update_item_availablity in route "/api/equpiment/update-item"
which sets an `item` (searched for by it's id) to the `availability`
parameter, which is set to True by default.

This action requires the `eqiupment.hide` permission. I also placed the
api route in the equipment_reservation file as we expect more than just
admins will be using this route.