Adding Tests and Item Routes (#51)
* Add and Delete Operating Hours from Coworking Database via API (#136)

* Fix some typing issues in backend/main.py
* Add API for listing Coworking Operating Hours
* Enforce permissions on OperatingHours Create
* Delete OperatingHours via the API

* Add Code File Labels and some Clarification to Event Binding Section (#139)

* Add Clarification to Event Binding (#140)

* Improve documentation of database and branching practices (#138)

* Document Database concerns
* Document Branching Practices and PR/CR Strategies

* Add Extra Resource to Further Reading Section (#141)

* Finalize notes on Merging a PR via Squash and Merge

* Add PR Template to the Docs Page (#142)

This pull request adds a template for PRs to the `branches` docs page.

* Update widgets.md (#143)

* Add Clarification Pt based on a student’s feedback (#144)

* Make Organization, Event, and About Page Public (#146)

* Feature Permissions Document (#149)

This commit adds a document to the docs directory that defines the
standard permissions used in the backend to authorize users for feature
actions. This ensures that permissions are added properly in the
frontend by providing a reference file for existing action/resource
sets.

* Resolve Date Time Issues (#150)

* Completed Equipment Type and Item's Table in Databse (#38)

* Created Entity Files for Types and Items

Added type_id field to EquipmentItem model to support the foreign key
in the equipment-items table. Created to_model, to_details_model, and
from_model methods on each entity object to support translations.

* Completed get_all_types Full Route Functionality

Deffered service method to database to collect all types. Also created
demo data to be inserted into databse. Further, created test to test
get_all_types() functionality. The test relies on the imports and on
fixtures.py and core_data.py as well as the demo data.

* Equipment type backend routes (#40)

* Refactor equipment api and add admin api file

The equipment.py file that existed previously has been renamed
and moved into a new directory to align with the specifications
of the design doc's apis. equipment_admin has been started to
include routes for adding, modifying, and deleting equipment
types.

* Begin skeleton route implementation

* get_all and get_items_from_type

Also created tests and demo data following the conventions set forth
by the organizations module.

* Equipment type backend routes (#46)

* Refactor equipment api and add admin api file

The equipment.py file that existed previously has been renamed
and moved into a new directory to align with the specifications
of the design doc's apis. equipment_admin has been started to
include routes for adding, modifying, and deleting equipment
types.

* Begin skeleton route implementation

* Add routes and service layer for add, edit, delete

The routes and the service layer for the admin to add, modify,
and delete equipment types have been added and connected together.

* Update documentation and integrate PR improvements

The documentation for the parameters/args of service methods
and api routes have begun to be updated and suggested changes
to permission strings and return statements.

* Finish args documentation update equipment service

* Added get_items_from_type route

* Added Null id check

* Completed Tests for Type Methods

Modified tablenames of EquipmentItem and EquipmentType to support
external operations as it seems that the dash "-" was causing errors.
All tests pass except for delete_type which I suspect is due to deleting
items without deleting their related items first. Will be fixing in the
next commit.

* Added Service item add/remove, Fixed Delete

Delete type also needs to delete all it's associated items for an error
to not be thrown.

* Updated Tests for 100% Coverage

Fixed potential error in delete_type since the check for null entity
was being run after attempting to delete all items of that type.

* Completed API Updates

Added /create-item and /delete-item routes. Updated routes to support
best practice HTTP Exception throwing. This will allow for our
documetation to be clearer, as well as easier error handling on the
frontend.

* Added update_item_availablity

Created update_item_availablity in route "/api/equpiment/update-item"
which sets an `item` (searched for by it's id) to the `availability`
parameter, which is set to True by default.

This action requires the `eqiupment.hide` permission. I also placed the
api route in the equipment_reservation file as we expect more than just
admins will be using this route.

* Fix Agreement

* Edited modify-type to get rid of redundant id field

---------

Co-authored-by: Kris Jordan <kris@cs.unc.edu>
Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>
Co-authored-by: Jade Keegan <97476936+jadekeegan@users.noreply.github.com>
Co-authored-by: Benjamin Eldridge <beldridge@unc.edu>
Co-authored-by: Benjamin Eldridge <97560111+benjabenn@users.noreply.github.com>