Fix Incorrectly Named Model & Create Sample Data
Created ItemDetails, EquipmentReservation, and EquipmentType models.
Changed the old EquipmentType model to instead be named ItemDetails to
better reflect the convetions set up in the backend and to prevent
circular types in the frontend.

Also experimented with using resolvers for the calendar view but
realized the API call to get the list of ItemDetails will suffice as
each item will have an equipment_type field and we can get the values
from there.