Complete Calendar Frontend Only
Created calendar component that currenlty only works with dummy data
supplied in the equipment service. The calendar will add items as they
are selected to lists, then when the submit button is pressed, the
selected lists are checked to make sure only 1 item was selected and the
dates were consecuitive.

The CSS for the calendar was the toughest part, I was unable to figure
out why the red squares are slightly smaller than the clickable squares.
I also had difficulty creating the CSS for the key and how-to sections.