Fix date selection and Validation
Had an issue with the .indexOf() method not finding the correct index
while using Date objects. Instead switched to storing dates as strings,
then converting to Date objects in the onSubmit() method.

Also ran into an issue with converting from a string to a date since
that assumes a timezone of GMT+0, so changed to using the number based
constructor which creates for current timezone at midnight.