Created ambassador equipment view
This commit creates the view for the ambassador to check in/out
equipment and view past reservations. An ambassador home component
and service were created, as well as an equipment reservation model.

Some small backend changes were made as well, such as to the route
for returning reservations. The date is now calculated on the backend
and the return description is passed in on the URL instead of the date.