Fix EquipmentDisplayComponent Bugs
Had a bug stating that `date.getTime()` was not a function, found
that creating a new date from the old date was enough to fix this
problem.

Also had a bug that the `reservation.type` was undefined. Found that
this field's name was changed to `equipment_type` so I changed that
on the frontend as well.