Ambassador view frontend (#70)
* Created Basic Reservation Model and Entity

* Update equipment_reservation.py routes

* Add parameter types to routes

* Create ambassador home component and service

* Add permission checking for /ambassador access

The permission "equipment.reservation.*" for resource "equipment"
has been added as necessary for access to the equipment reservation
ambassador URL. Additionally, I have added a frontend model for
reservations.

* Model and entity complete

* Create api and service file for reservations

* Add reservation schedule api to main

* Move to_details() to service layer to avoid circular dependency between Item and Reservation Entities

* Complete create_reservation and test

* Add get_user_equipment_reservations and activate_reservations routes

* Update EOL

* Complete all reservation routes

* Clean comments and imports.

* Clean comments and imports.

* Clean comments and imports.

* Add relationships to EquipmentReservationEntities

* Complete create_reservation and checking for availability.

* Complete create_reservation and checking for availability.

* Solve review comments.

* Solve review comments.

* Solve review comments.

* Created ambassador equipment view

This commit creates the view for the ambassador to check in/out
equipment and view past reservations. An ambassador home component
and service were created, as well as an equipment reservation model.

Some small backend changes were made as well, such as to the route
for returning reservations. The date is now calculated on the backend
and the return description is passed in on the URL instead of the date.

* Add newline for return description

This commit ensures that each return description "entry" ends with a
newline character as added in the component.

* Fixed equipment_type name error

Additionally, a console.log call was removed.

* Add new return description functionality

* Store return description in item_entity

* Update documentation for model change

---------

Co-authored-by: Andrew Lockard <alockard@unc.edu>
Co-authored-by: Jacob Fessler <jacobdan@ad.unc.edu>
Co-authored-by: VitorInserra <inserra.vitor@gmail.com>