Student Reserve Screen (#69)
* Setup

* Created Basic Reservation Model and Entity

* Update equipment_reservation.py routes

* Add parameter types to routes

* Model and entity complete

* Create api and service file for reservations

* Add reservation schedule api to main

* Completed Calendar Square Widget

Takes in a date and a reserved status, the date will be emitted when
the calendar square is selected, which will be used in the component to
store dates as they are clicked.

* Modifited Equipment Routing Module

Following the convetion of including a components route information
in the componet as a static field called "route", then including this
field in the routing module.

* Move to_details() to service layer to avoid circular dependency between Item and Reservation Entities

* Complete create_reservation and test

* Add get_user_equipment_reservations and activate_reservations routes

* Update EOL

* Updated Equipment Card widget

Made it so the 'Check-Out' button will be greyed out when there are no
items available for checkout. Also added link to correct reserve screen.

* Complete all reservation routes

* Fix Incorrectly Named Model & Create Sample Data

Created ItemDetails, EquipmentReservation, and EquipmentType models.
Changed the old EquipmentType model to instead be named ItemDetails to
better reflect the convetions set up in the backend and to prevent
circular types in the frontend.

Also experimented with using resolvers for the calendar view but
realized the API call to get the list of ItemDetails will suffice as
each item will have an equipment_type field and we can get the values
from there.

* Clean comments and imports.

* Clean comments and imports.

* Clean comments and imports.

* Add relationships to EquipmentReservationEntities

* Complete create_reservation and checking for availability.

* Complete create_reservation and checking for availability.

* Complete Calendar Frontend Only

Created calendar component that currenlty only works with dummy data
supplied in the equipment service. The calendar will add items as they
are selected to lists, then when the submit button is pressed, the
selected lists are checked to make sure only 1 item was selected and the
dates were consecuitive.

The CSS for the calendar was the toughest part, I was unable to figure
out why the red squares are slightly smaller than the clickable squares.
I also had difficulty creating the CSS for the key and how-to sections.

* Solve review comments.

* Solve review comments.

* Solve review comments.

* Changed to accept data via backend routes

Switched to using a resolver so the items list can be preloaded into
the ReserveScreenComponent before construction. Also added check for
max reservation time.

* Fix date selection and Validation

Had an issue with the .indexOf() method not finding the correct index
while using Date objects. Instead switched to storing dates as strings,
then converting to Date objects in the onSubmit() method.

Also ran into an issue with converting from a string to a date since
that assumes a timezone of GMT+0, so changed to using the number based
constructor which creates for current timezone at midnight.

* Connect to backend create reservation API

* Added Reservation View to main Equipment Page

Note: I have not tested this for actions requiring ambassadors.

* CreateReservation Error and Redirect Handling

* Convert to Jacob's Styling Suggestion

* Changed Available to Items

This wording is better suited to what the underlying code is doing to
calculate this number. We are not calculating the total amount that the
user can currently reserve, but instead we are calculating the total
number of items that a user can currently see on the reserve screen.

* Update Calendar Display Width for 10 Days

* Add Error parsing for creating reservations

* Fix EquipmentDisplayComponent Bugs

Had a bug stating that `date.getTime()` was not a function, found
that creating a new date from the old date was enough to fix this
problem.

Also had a bug that the `reservation.type` was undefined. Found that
this field's name was changed to `equipment_type` so I changed that
on the frontend as well.

* Modify How-To

* Fixed Date Pipe Consistency

* Fix display issues on Past Reservations

---------

Co-authored-by: Jacob Fessler <jacobdan@ad.unc.edu>
Co-authored-by: VitorInserra <inserra.vitor@gmail.com>